+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/controller/BindingController.groovy
@@ -30,12 +30,12 @@ import com.swisscom.cloud.sb.broker.repository.CFServiceRepository
 import com.swisscom.cloud.sb.broker.repository.PlanRepository
 import com.swisscom.cloud.sb.broker.repository.ServiceBindingRepository
 import com.swisscom.cloud.sb.broker.services.common.ServiceProvider
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
-import com.swisscom.cloud.sb.broker.services.credential.BindResponse
-import com.swisscom.cloud.sb.broker.services.credential.FetchServiceBindingProvider
-import com.swisscom.cloud.sb.broker.services.credential.ServiceBindingPersistenceService
-import com.swisscom.cloud.sb.broker.services.credential.ServiceInstanceBindingResponseDto
-import com.swisscom.cloud.sb.broker.services.credential.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
+import com.swisscom.cloud.sb.broker.binding.BindResponse
+import com.swisscom.cloud.sb.broker.binding.FetchServiceBindingProvider
+import com.swisscom.cloud.sb.broker.binding.ServiceBindingPersistenceService
+import com.swisscom.cloud.sb.broker.binding.ServiceInstanceBindingResponseDto
+import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.util.Audit
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
