+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/services/credhub/CredHubResponseDto.groovy
@@ -1,6 +1,6 @@
 package com.swisscom.cloud.sb.broker.services.credhub
 
-import com.swisscom.cloud.sb.broker.services.credential.BindResponseDto
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import groovy.json.JsonBuilder
 
 class CredHubResponseDto implements BindResponseDto{
