+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/services/kubernetes/dto/RedisBindResponseDto.groovy
@@ -15,7 +15,7 @@
 
 package com.swisscom.cloud.sb.broker.services.kubernetes.dto
 
-import com.swisscom.cloud.sb.broker.services.credential.BindResponseDto
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import groovy.json.JsonBuilder
 
 class RedisBindResponseDto implements BindResponseDto {
