+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/services/relationaldb/RelationalDbServiceProvider.groovy
@@ -15,9 +15,9 @@
 
 package com.swisscom.cloud.sb.broker.services.relationaldb
 
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
-import com.swisscom.cloud.sb.broker.services.credential.BindResponse
-import com.swisscom.cloud.sb.broker.services.credential.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
+import com.swisscom.cloud.sb.broker.binding.BindResponse
+import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.context.CloudFoundryContextRestrictedOnly
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
