+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummyExtension/DummyExtensionsServiceProvider.groovy
@@ -15,9 +15,9 @@
 
 package com.swisscom.cloud.sb.broker.util.test.DummyExtension
 
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
-import com.swisscom.cloud.sb.broker.services.credential.BindResponse
-import com.swisscom.cloud.sb.broker.services.credential.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
+import com.swisscom.cloud.sb.broker.binding.BindResponse
+import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
 import com.swisscom.cloud.sb.broker.model.UpdateRequest
