+++ broker/core/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummyServiceProvider.groovy
@@ -20,10 +20,10 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.common.base.Optional
 import com.google.gson.Gson
 import com.swisscom.cloud.sb.broker.async.AsyncProvisioningService
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
-import com.swisscom.cloud.sb.broker.services.credential.BindResponse
-import com.swisscom.cloud.sb.broker.services.credential.BindResponseDto
-import com.swisscom.cloud.sb.broker.services.credential.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
+import com.swisscom.cloud.sb.broker.binding.BindResponse
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
+import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointProvider
 import com.swisscom.cloud.sb.broker.model.*
 import com.swisscom.cloud.sb.broker.provisioning.DeprovisionResponse
