+++ broker/core/src/test/groovy/com/swisscom/cloud/sb/broker/services/credhub/CredHubServiceProviderSpec.groovy
@@ -1,6 +1,6 @@
 package com.swisscom.cloud.sb.broker.services.credhub
 
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
 import com.swisscom.cloud.sb.broker.error.ServiceBrokerException
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.Plan
