+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/mongodb/MongoDbBindResponseDto.groovy
@@ -15,7 +15,7 @@
 
 package com.swisscom.cloud.sb.broker.services.mongodb
 
-import com.swisscom.cloud.sb.broker.services.credential.BindResponseDto
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import groovy.json.JsonBuilder
 
 class MongoDbBindResponseDto implements BindResponseDto {
