+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/openwhisk/OpenWhiskBindResponseDto.groovy
@@ -15,7 +15,7 @@
 
 package com.swisscom.cloud.sb.broker.services.openwhisk
 
-import com.swisscom.cloud.sb.broker.services.credential.BindResponseDto
+import com.swisscom.cloud.sb.broker.binding.BindResponseDto
 import groovy.json.JsonBuilder
 
 class OpenWhiskBindResponseDto implements BindResponseDto {
