+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/services/mariadb/MariaDBServiceProviderSpec.groovy
@@ -15,8 +15,8 @@
 
 package com.swisscom.cloud.sb.broker.services.mariadb
 
-import com.swisscom.cloud.sb.broker.services.credential.BindRequest
-import com.swisscom.cloud.sb.broker.services.credential.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.BindRequest
+import com.swisscom.cloud.sb.broker.binding.UnbindRequest
 import com.swisscom.cloud.sb.broker.error.ServiceBrokerException
 import com.swisscom.cloud.sb.broker.model.*
 import com.swisscom.cloud.sb.broker.repository.ServiceInstanceRepository
