+++ model/src/main/groovy/com/swisscom/cloud/sb/broker/binding/BindRequest.groovy
@@ -13,14 +13,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.swisscom.cloud.sb.broker.services.credential
+package com.swisscom.cloud.sb.broker.binding
 
-class AbstractBindResponseDto implements BindResponseDto {
-    String syslog_drain_url
-    String route_service_url
+import com.swisscom.cloud.sb.broker.model.CFService
+import com.swisscom.cloud.sb.broker.model.Plan
+import com.swisscom.cloud.sb.broker.model.ServiceInstance
 
-    @Override
-    String toJson() {
-        return null
-    }
+class BindRequest {
+    ServiceInstance serviceInstance
+    String binding_guid
+    String app_guid
+    Plan plan
+    CFService service
+    Map parameters
 }
