+++ model/src/main/groovy/com/swisscom/cloud/sb/broker/binding/BindResponse.groovy
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.swisscom.cloud.sb.broker.services.credential
+package com.swisscom.cloud.sb.broker.binding
 
 import com.swisscom.cloud.sb.broker.model.ServiceDetail
