+++ model/src/main/groovy/com/swisscom/cloud/sb/broker/binding/BindResponseDto.groovy
@@ -13,17 +13,8 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.swisscom.cloud.sb.broker.services.credential
+package com.swisscom.cloud.sb.broker.binding
 
-import com.swisscom.cloud.sb.broker.model.CFService
-import com.swisscom.cloud.sb.broker.model.Plan
-import com.swisscom.cloud.sb.broker.model.ServiceInstance
-
-class BindRequest {
-    ServiceInstance serviceInstance
-    String binding_guid
-    String app_guid
-    Plan plan
-    CFService service
-    Map parameters
+interface BindResponseDto extends Serializable {
+    String toJson()
 }
